requirements.txt
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dash==1.4.0
dash-bootstrap-components==0.7.1
dash-core-components==1.3.0
dash-daq==0.2.1
dash-html-components==1.0.1
dash-renderer==1.1.1
dash-table==4.4.0
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.1
plotly==4.1.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
